chore: move stroke events and image events into separate files
