+++ Scribble.Shared/Lib/Event.cs
@@ -54,41 +54,6 @@ public interface ITerminalEvent
 {
 }
 
-public record StartStrokeEvent(
-    Guid ActionId,
-    Guid StrokeId,
-    SKPoint StartPoint,
-    StrokePaint StrokePaint,
-    ToolType ToolType,
-    HashSet<ToolOption> ToolOptions)
-    : Event(ActionId);
-
-public record EndStrokeEvent(Guid ActionId) : Event(ActionId), ITerminalEvent;
-
-// PENCIL TOOL
-public record PencilStrokeLineToEvent(Guid ActionId, Guid StrokeId, SKPoint Point) : Event(ActionId);
-
-// LINE + ARROW + RECTANGLE TOOL
-public record LineStrokeLineToEvent(Guid ActionId, Guid StrokeId, SKPoint EndPoint) : Event(ActionId);
-
-// TEXT TOOL
-public record AddTextEvent(
-    Guid ActionId,
-    Guid StrokeId,
-    SKPoint Position,
-    string Text,
-    StrokePaint Paint,
-    HashSet<ToolOption> ToolOptions)
-    : Event(ActionId), ITerminalEvent;
-
-// IMAGE TOOL
-public record AddImageEvent(
-    Guid ActionId,
-    Guid ImageId,
-    string ImageBase64String,
-    SKPoint Position,
-    SKSize Size) : Event(ActionId), ITerminalEvent;
-
 // MISC
 public record LoadCanvasEvent(Guid ActionId, List<CanvasElement> CanvasElements) : Event(ActionId);
