+++ Scribble/Services/CanvasStateService/Handlers/StrokeReplayHandler.cs
@@ -2,6 +2,7 @@ using System;
 using Scribble.Services.CanvasStateService.State;
 using Scribble.Shared.Lib;
 using Scribble.Shared.Lib.CanvasElements.Strokes;
+using Scribble.Shared.Lib.Events;
 using Scribble.Tools.PointerTools.ArrowTool;
 using Scribble.Utils;
 using SkiaSharp;
