+++ Scribble/Tools/PointerTools/ImageTool/ImageTool.cs
@@ -9,7 +9,7 @@ using Avalonia.Threading;
 using Scribble.Services.CanvasStateService;
 using Scribble.Services.DialogService;
 using Scribble.Services.FileService;
-using Scribble.Shared.Lib;
+using Scribble.Shared.Lib.Events;
 using SkiaSharp;
 
 namespace Scribble.Tools.PointerTools.ImageTool;
